[EMBR-6026] Avoid code duplication when starting native SDK (cleanup)
