+++ packages/core/.gitignore
@@ -1,1 +1,3 @@
+# Custom
 src/interfaces/common
+ios/common
