+++ packages/react-native-otlp/.gitignore
@@ -75,3 +75,4 @@ yarn-error.log
 
 # Custom
 src/interfaces/common
+ios/common
