+++ packages/react-native-otlp/package.json
@@ -18,7 +18,9 @@
     "test": "__tests__"
   },
   "scripts": {
-    "build": "cp -r ../../interfaces/ ./src/interfaces/common && tsc",
+    "ios:shared": "mkdir -p ./ios/common && cp -r ../common/ios/* ./ios/common/",
+    "js:shared": "mkdir -p ./src/interfaces/common && cp -r ../common/interfaces/* ./src/interfaces/common",
+    "build": "yarn run js:shared && yarn run ios:shared && tsc",
     "android:test": "cd test-project && yarn && cd android && ./gradlew :app:detek && ./gradlew :app:testDebugUnitTest && cd ../../",
     "ios:test": "../../scripts/run-ios-tests.sh test-project/ios/RNEmbraceOTLPTests.xcworkspace RNEmbraceOTLPTests",
     "ios:install": "cd test-project && yarn install && cd ios && pod install && rm .xcode.env.local"
